sndfile: fix precision

+++ plugins/sndfile/sndfile.c
@@ -359,7 +359,7 @@ sndfile_seek_sample (DB_fileinfo_t *_info, int sample) {
 
 static int
 sndfile_seek (DB_fileinfo_t *_info, float sec) {
-    return sndfile_seek_sample64 (_info, sec * _info->fmt.samplerate);
+    return sndfile_seek_sample64 (_info, (int64_t)((double)sec * (int64_t)_info->fmt.samplerate));
 }
 
 static int
